requirements.txt
-i https://pypi.org/simple

anyio==3.6.1; python_full_version >= '3.6.2'

bcrypt==4.0.0

cffi==1.15.1

click==8.1.3; python_version >= '3.7'

colorama==0.4.5

cryptography==37.0.4

ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

fastapi==0.81.0

h11==0.13.0; python_version >= '3.6'

httptools==0.4.0

idna==3.3; python_version >= '3.5'

passlib[bcrypt]==1.7.4

peewee==3.15.2

psycopg2==2.9.3

pyasn1==0.4.8

pycparser==2.21

pydantic==1.10.1; python_version >= '3.7'

python-dotenv==0.20.0

python-jose[cryptography]==3.3.0

python-multipart==0.0.5

pyyaml==6.0

rsa==4.9; python_version >= '3.6' and python_version < '4'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sniffio==1.2.0; python_version >= '3.5'

starlette==0.19.1; python_version >= '3.6'

typing-extensions==4.3.0; python_version >= '3.7'

uvicorn[standard]==0.18.3

watchfiles==0.16.1

websockets==10.3
pytest~=7.1.2
pydantic~=1.10.1